If the image is no need to be cached, I would like to mark it as cached image.
For example, local image file is no need to be cached.
If all thread of taskExecutor is running for slow network tasks.
Local image can not be downloaded until those slow network tasks
are finished.

But cached images can be downloaded by taskExecutorForCachedImages
in the same time. The download speed of local image is as
fast as cached image. So I would like to mark it as cached image.